requirements.txt
absl-py==2.1.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
chatbot==1.5.2b0
chatterbot-corpus==1.2.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
crayons==0.4.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
Django==5.0.6
dnspython==2.6.1
executing==2.0.1
filterpy==1.4.5
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.56.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.67.0
grpcio-tools==1.67.0
h5py==3.12.1
idna==3.10
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mysql-connector-python==9.2.0
namex==0.0.8
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.0
paho-mqtt==2.1.0
pandas==2.2.1
parso==0.8.4
piexif==1.1.3
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
protobuf==5.28.2
psutil==5.9.8
pure-eval==0.2.2
pydub==0.25.1
pygame==2.5.2
Pygments==2.18.0
pymongo==4.8.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
pywin32==306
PyYAML==3.13
pyzbar==0.1.9
pyzmq==26.0.3
qrcode==8.0
regex==2023.12.25
requests==2.32.3
rich==13.9.4
scipy==1.15.2
setuptools==69.2.0
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
stack-data==0.6.3
stegano==0.11.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tkinterdnd2==0.3.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.43.0
wrapt==1.16.0
